리뷰Context분리, 배포용PATH적용, npm install 하나 필요, 리뷰삭제 에러로직수정_12.23 17:41_성찬

+++ src/main/java/com/example/demo/store/controller/ReviewController.java
@@ -100,8 +100,14 @@ public class ReviewController {
 
     // 리뷰 삭제
     @DeleteMapping("/{reviewId}")
-    public ResponseEntity<String> deleteReview(@PathVariable int reviewId) {
+    public ResponseEntity<String> deleteReview(
+            @PathVariable(name = "reviewId") int reviewId,
+            @RequestHeader(value = "Authorization", required = true) String token) {
         try {
+            if (!token.startsWith("Bearer ")) {
+                return ResponseEntity.status(HttpStatus.SC_UNAUTHORIZED).body("유효하지 않은 토큰입니다.");
+            }
+            
             reviewService.deleteReview(reviewId);
             return ResponseEntity.ok("리뷰가 성공적으로 삭제되었습니다.");
         } catch (RuntimeException e) {
